Use CMAKE_INSTALL_FULL_* for absolute paths
Instead of using a combination of CMAKE_INSTALL_PREFIX and
CMAKE_INSTALL_, we should use CMAKE_INSTALL_FULL_.

If the CMAKE_INSTALL_ is an absolute path (e.g. in NixOS), then
prefixing it with another absolute path would be incorrect.
CMAKE_INSTALL_FULL_ already handles the case if
CMAKE_INSTALL_ is an absolute path. See:
https://cmake.org/cmake/help/latest/module/GNUInstallDirs.html

Signed-off-by: Daniel Fullmer <dfullmer@anduril.com>